requirements.txt
RPi.GPIO==0.7.0
pyserial==3.4
